requirements.txt
fastapi==0.115.9
pydantic==2.11.4
uvicorn==0.34.2
python-multipart==0.0.20
torch==2.7.0
transformers==4.51.3
sentence-transformers==4.1.0
langchain==0.3.25
langchain-community==0.3.23
faiss-cpu==1.11.0
pypdf==5.4.0
docx2txt==0.9
qdrant-client==1.14.2
dill==0.4.0
langchain_openai==0.3.16
unstructured==0.17.2
sentence-transformers==4.1.0
PyPDF2==3.0.1
python-multipart==0.0.20
python-dotenv==1.1.0